requirements.txt
google-auth
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
sqlalchemy
pdfplumber
pandas
python-dotenv
mock
pytest
imapclient
pyzmail36
python-dotenv
psycopg2-binary
google-cloud-pubsub